implemented the bull
