+++ app/backend/src/jobs/jobs.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { BullModule } from '@nestjs/bullmq';
 import { JobsController } from './jobs.controller';
+import { JobsMonitorService } from './jobs-monitor.service';
 
 @Module({
   imports: [
@@ -9,5 +10,7 @@ import { JobsController } from './jobs.controller';
     BullModule.registerQueue({ name: 'onchain' }),
   ],
   controllers: [JobsController],
+  providers: [JobsMonitorService],
+  exports: [JobsMonitorService],
 })
 export class JobsModule {}
